requirements.txt
yfinance
pandas_datareader
fredapi
vaderSentiment
scikit-learn
streamlit
plotly
matplotlib
seaborn
